refactor: Add the apm internal package
