+++ internal/status/api/v1/manager.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"reflect"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -227,51 +226,6 @@ func (m *StatusManager) rebuildPolicyStatuses() error {
 	return nil
 }
 
-func (m *StatusManager) updateArmorProfileStatus(
-	ap *varmor.ArmorProfile,
-	policyStatus *varmortypes.PolicyStatus) error {
-
-	var conditions []varmor.ArmorProfileCondition
-	for nodeName, message := range policyStatus.NodeMessages {
-		if message != string(varmortypes.ArmorProfileReady) {
-			c := newArmorProfileCondition(nodeName, varmortypes.ArmorProfileReady, v1.ConditionFalse, "", message)
-			conditions = append(conditions, *c)
-		}
-	}
-
-	regain := false
-	update := func() (err error) {
-		if regain {
-			ap, err = m.varmorInterface.ArmorProfiles(ap.Namespace).Get(context.Background(), ap.Name, metav1.GetOptions{})
-			if err != nil {
-				if k8errors.IsNotFound(err) {
-					return nil
-				}
-				return err
-			}
-		}
-		// Nothing needs to be updated.
-		if reflect.DeepEqual(ap.Status.Conditions, conditions) &&
-			ap.Status.CurrentNumberLoaded == policyStatus.SuccessedNumber &&
-			ap.Status.DesiredNumberLoaded == m.desiredNumber {
-			return nil
-		}
-		ap.Status.DesiredNumberLoaded = m.desiredNumber
-		ap.Status.CurrentNumberLoaded = policyStatus.SuccessedNumber
-		if len(conditions) > 0 {
-			ap.Status.Conditions = conditions
-		} else {
-			ap.Status.Conditions = nil
-		}
-		_, err = m.varmorInterface.ArmorProfiles(ap.Namespace).UpdateStatus(context.Background(), ap, metav1.UpdateOptions{})
-		if err != nil {
-			regain = true
-		}
-		return err
-	}
-	return retry.RetryOnConflict(retry.DefaultRetry, update)
-}
-
 func (m *StatusManager) updateVarmorPolicyStatus(
 	vp *varmor.VarmorPolicy,
 	ready bool,
@@ -442,9 +396,9 @@ func (m *StatusManager) reconcileStatus(stopCh <-chan struct{}) {
 				logger.Error(err, "m.varmorInterface.ArmorProfiles().Get()")
 				break
 			}
-			err = m.updateArmorProfileStatus(ap, &policyStatus)
+			err = UpdateArmorProfileStatus(m.varmorInterface, ap, &policyStatus, m.desiredNumber)
 			if err != nil {
-				logger.Error(err, "m.updateArmorProfileStatus()")
+				logger.Error(err, "UpdateArmorProfileStatus()")
 				break
 			}
 
@@ -567,7 +521,7 @@ func (m *StatusManager) reconcileStatus(stopCh <-chan struct{}) {
 			}
 
 			apName := varmorprofile.GenerateArmorProfileName(namespace, vpName, clusterScope)
-			profile, err := varmorprofile.GenerateProfile(vPolicy, apName, namespace, m.varmorInterface, true)
+			profile, err := varmorprofile.GenerateProfile(vPolicy, apName, namespace, m.varmorInterface, true, logger)
 			if err != nil {
 				logger.Error(err, "varmorprofile.GenerateProfile()")
 				break
